requirements.txt
ultralytics
cvzone
opencv-python
telegram
requests
cryptography
python-dotenv
pytest

